+++ services/chaindb/postgresql/attestations.go
@@ -21,7 +21,7 @@ import (
 	"strings"
 
 	"github.com/attestantio/go-eth2-client/spec/phase0"
-	"github.com/jackc/pgx/v4"
+	"github.com/jackc/pgx/v5"
 	"github.com/pkg/errors"
 	"github.com/wealdtech/chaind/services/chaindb"
 	"go.opentelemetry.io/otel"