+++ services/chaindb/postgresql/metadata.go
@@ -16,8 +16,7 @@ package postgresql
 import (
 	"context"
 
-	"github.com/jackc/pgtype"
-	"github.com/jackc/pgx/v4"
+	"github.com/jackc/pgx/v5"
 	"github.com/pkg/errors"
 	"go.opentelemetry.io/otel"
 )
@@ -63,13 +62,13 @@ func (s *Service) Metadata(ctx context.Context, key string) ([]byte, error) {
 		defer s.CommitROTx(ctx)
 	}
 
-	res := &pgtype.JSONB{}
+	res := make([]byte, 0)
 	err = tx.QueryRow(ctx, `
       SELECT f_value
       FROM t_metadata
       WHERE f_key = $1`,
 		key).Scan(
-		res,
+		&res,
 	)
 
 	if err != nil {
@@ -79,5 +78,5 @@ func (s *Service) Metadata(ctx context.Context, key string) ([]byte, error) {
 		return nil, errors.Wrap(err, "failed to obtain metadata")
 	}
 
-	return res.Bytes, nil
+	return res, nil
 }