+++ main.go
@@ -61,7 +61,7 @@ import (
 )
 
 // ReleaseVersion is the release version for the code.
-var ReleaseVersion = "0.7.7-dev"
+var ReleaseVersion = "0.7.7"
 
 func main() {
 	os.Exit(main2())